[Refactor] : 스티커 박스 컴포넌트 분리
useUploadSticker 훅을 이용해 input을 이용해 스티커 업로드하는 로직을 분리(관심사의 분리)
